HtmlEditor: Move tests to the separate matrix jobs. (#23217)
Cherry-pick of the 3c7d15f654a80eb28878eb567e532b9061e28a3e.
