+++ testing/testcafe/tests/htmlEditor/etalons/htmleditor-with-toolbar-focused.png
