+++ testing/testcafe/tests/htmlEditor/etalons/htmleditor-without-toolbar.png
