+++ testing/tests/DevExpress.serverSide.htmlEditor/__meta.json
@@ -1,0 +1,5 @@
+{
+    "constellation": "ui.htmlEditor",
+    "explicit": false,
+    "runOnDevices": false
+}
