+++ testing/tests/DevExpress.ui.widgets.htmlEditor/htmlEditorParts/dropImageModule.tests.js
