+++ testing/tests/DevExpress.ui.widgets.htmlEditor/htmlEditorParts/imageCursorModule.tests.js
