+++ testing/tests/DevExpress.ui.widgets.htmlEditor/htmlEditorParts/imageUploadModule.tests.js
