+++ testing/tests/DevExpress.ui.widgets.htmlEditor/htmlEditorParts/paste.tests.js
