+++ testing/tests/DevExpress.ui.widgets.htmlEditor/htmlEditorParts/quillRegistrator.tests.js
