+++ testing/tests/DevExpress.ui.widgets.htmlEditor/htmlEditorParts/tableResizingIntegration.tests.js
