+++ testing/tests/DevExpress.ui.widgets.htmlEditor/htmlEditorParts/utils.js
