+++ testing/tests/DevExpress.ui.widgets.htmlEditor/htmlEditorParts/variablesModule.tests.js
